refactor(profile): move code
